# crearConnector/ConnectorASDKV8/ConectorServiceDesk3.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ConectorServiceDesk4",
    "description": "",
    "version": "1.0"
  },
  "host": "hostUrl",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/api/ManagmentParameters/Services/{projectId}": {
      "get": {
        "responses": {
          "200": {
            "description": "default",
            "schema": {
              "type": "object",
              "properties": {
                "parameters": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "format": "int32",
                        "description": "id"
                      },
                      "name": {
                        "type": "string",
                        "description": "name"
                      },
                      "type": {
                        "type": "integer",
                        "description": "type",
                        "title": "",
                        "format": "int32"
                      }
                    }
                  },
                  "description": "parameters"
                }
              }
            }
          }
        },
        "operationId": "GetServices",
        "summary": "Trae los servicios relacionados con el usuario y el proyecto seleccionado",
        "description": "Trae los servicios relacionados con el usuario y el proyecto seleccionado",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "x-ms-dynamic-values": {
              "operationId": "Proyect",
              "value-path": "id",
              "value-collection": "parameters",
              "value-title": "name"
            }
          }
        ],
        "x-ms-visibility": "internal"
      }
    },
...
    "/api/ManagmentParameters/State/{itemType}/{projectId}": {
      "get": {
        "responses": {
          "200": {
            "description": "default",
            "schema": {
              "type": "object",
              "properties": {
                "parameters": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "format": "int32",
                        "description": "id"
                      },
                      "name": {
                        "type": "string",
                        "description": "name"
                      },
                      "type": {
                        "type": "integer",
                        "format": "int32",
                        "description": "type"
                      }
                    }
                  },
                  "description": "parameters"
                }
              }
            }
          }
        },
        "summary": "Obtener estados",
        "description": "Obtener estados",
        "operationId": "GetState",
        "x-ms-visibility": "internal",
        "parameters": [
          {
            "name": "itemType",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "x-ms-dynamic-values": {
              "operationId": "GetItemType",
              "value-path": "id",
              "value-collection": "parameters",
              "value-title": "name",
              "parameters": {
                "categoryId": {
                  "parameter": "itemType"
                },
                "projectId": {
                  "parameter": "projectId"
                }
              }
            }
          },
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "x-ms-dynamic-values": {
              "operationId": "Proyect",
              "value-path": "id",
              "value-collection": "parameters",
              "value-title": "name"
            }
          }
        ]
      }
    },
...
    "/api/ManagmentParameters/Responsible/{groupId}/{projectId}": {
      "get": {
        "responses": {
          "default": {
            "description": "default",
            "schema": {
              "type": "object",
              "properties": {
                "parameters": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "format": "int32",
                        "description": "id"
                      },
                      "name": {
                        "type": "string",
                        "description": "name"
                      }
                    }
                  },
                  "description": "parameters"
                }
              }
            }
          }
        },
        "summary": "Responsable",
        "x-ms-visibility": "internal",
        "operationId": "GetResponsible",
        "description": "Responsable",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ]
      }
    },
...
  },
...
  "securityDefinitions": {
    "basic_auth": {
      "type": "basic"
    }
  },
  "security": [
    {
      "basic_auth": []
    }
  ],
...
}